Move Course Notifications to common/courses
