+++ common/appointment/decline.ts
@@ -4,8 +4,8 @@ import { getUser, getStudent, User, userForStudent } from '../user';
 import * as Notification from '../notification';
 import { getLogger } from '../logger/logger';
 import { getAppointmentForNotification } from './util';
-import { getNotificationContextForSubcourse } from '../mails/courses';
 import { RedundantError } from '../util/error';
+import { getNotificationContextForSubcourse } from '../courses/notifications';
 
 const logger = getLogger('Appointment');
